Python script version v1.1
- Added modularity
- Implement use of arguments [argparse]
- Record currently playing track
- README.md updated
- Enable access to read and modify playback state
- Use of Spotify / OAuth 2.0 authorization framework

It will record your currently playing track, if it is paused it will
resume at the current track position.
If Spotify URI is passed it will pause your spotify device while it
records the track once it finish it will resume the track again.
If option record is passed it will start recording what your computer is
playing.